There were bugs in how the start time of a simulation was handled. This had the effect that for example with the default t_save = 1 and dt = 0.1, the value that was stored in the t=1 slot was actually the value for t=0.1, the value in t=2 was the value for t=1.1 etc. The error is visible for example in Figure 17 in the vignette, where not much seems to be happening up to time 1.
In addition, if an effort array was specified, then the effort for t=1 was used to go from t=0 to t=0.1 and the effort for t=2 was used to go from t=0.1 to t=1.1.

The code was unnecessarily complicated by the fact that the effort array had a time dimension that was one shorter than the time dimension of n and n_pp. This was because the start time for the simulation was not taken from the effort array but was set to slightly earlier than the smallest time in the effort array, by an undocumented amount. Now the effort array has the same dimension as the simulation and the smallest time in the effort array is the start time (the time at which the initial condition is specified) and the largest time in the effort array is the largest time up to which the simulation is run.
The n-th effort entry is the effort used between the n-th time and the (n+1)st time. Thus the last effort entry is not actually used.

I have dropped the requirement that the end-time t_max is a multiple of dt. Instead the simulation runs until the largest time of the form t_min + n dt  that is no larger than t_max. So for example if dt=1.1 and t_max = 3 then the times used are 0, 1.1, 2.2
I added tests for the requirement that t_save is an integer multiple of dt.