Do not truncate predation kernel
The automatic truncation of the lognormal predation kernel at beta * exp(3 * sigma) is not such a good idea because even though the kernel is small there, the abundance of prey may be very high, leading to a visible artificial truncation in the consumed biomass spectrum.

Now there are two lognormal predation kernel functions: `lognormal_pred_kernel()` does not truncate and `truncated_lognormal_pred_kernel()` does truncate. The default is the untruncated kernel.

This loss of a cutoff means it is not longer possible to determine a default for min_w_pp by the criterium that the plankton should extend far enough down to cover the entire predation kernel of all fish. Instead there is now a fixed default. There is thus some nice code that has become irrelevant and has been removed.

Having a non-truncated kernel increases the problem of wrap-around in the fft evaluation of the predation rate, where the large number of small individuals can have a noticeable effect on the predation mortality on large individuals. To mitigate, I am now cutting of the predation rate from predator j at the largest size of predator j. To do this efficiently I have introduced a new slot `@ft_mask` in the MizerParams class.