Introduce external encounter rate
Now the model can include an external encounter rate that represents the rate at which a predator encounters food that is not explicitly modelled.
This encounter rate is set with `setExtEncounter()` or `ext_encounter<-()` and can be read with `getExtEncounter()` or `ext_encounter()`. It is stored in a new `ext_encounter` slot of the MizerParams class.